refactor(plugin-inbox): consolidate sync-pipeline machinery into connector SyncBinding
Address review feedback:
- Fold the pipeline SyncBinding service/layer/commit into the existing
  @dxos/plugin-connector `SyncBinding` namespace, so there is one `SyncBinding`
  and no import alias; delete the duplicate inbox module.
- Take `db` from `Database.Service` in the commit/layer/extract stage rather than
  threading a handle through the sync-op signatures.
- Inline `Effect.provide(SyncBinding.layer(...))` into each pipeline `pipe` chain
  (drop the intermediate `drain` binding).
